+++ app/src/main/java/com/taru/data/remote/weather/dto/inner/WeatherMainDto.kt
@@ -7,7 +7,7 @@ import com.squareup.moshi.JsonClass
  * Created by Niraj on 17-01-2023.
  */
 @JsonClass(generateAdapter = true)
-data class WeatherMainDto(val temp: Double,
+data class WeatherMainDto(val temp: Float,
                           val pressure: Int,
                           val humidity: Int,
                           @Json(name = "temp_min") val tempMin: Float,
